use react-query to fetch and cache data
